requirements.txt
Django==3.2
asgiref==3.7.2
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.6.2
djangorestframework==3.12.0
Pillow==9.5.0
django-filter==23.2
MarkupSafe==2.1.2 
certifi==2023.5.7 
cffi==1.15.1 
charset-normalizer==3.1.0 
coreapi==2.3.3 
coreschema==0.0.4 
cryptography==40.0.2 
defusedxml==0.7.1 
django-templated-mail==1.1.1 
djangorestframework-simplejwt==4.7.2 
djoser==2.1.0
idna==3.4 
itypes==1.2.0 
jinja2==3.1.2 
oauthlib==3.2.2 
pycparser==2.21 
pyjwt==2.7.0 
python3-openid==3.2.0 
requests==2.31.0 
requests-oauthlib==1.3.1 
six==1.16.0 
social-auth-app-django==4.0.0 
social-auth-core==4.4.2 
uritemplate==4.1.1 
urllib3==2.0.2
python-dotenv==1.0.0
